pyproject.toml
[project]
name = "bookdb"
dynamic = ["version"]
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Yousef Amirghofran", email = "yamirghofran@gmail.com" }
]
requires-python = ">=3.9,<3.14"
dependencies = [
    "numpy>=1.19", # 1.19 required by tensorflow 2.6
    "pandas>2,<3",
    "ipykernel>=4.6.1,<7",
    "jupyter>=1,<2",
    "zensical>=0.0.2",
    "fsspec>=2025.3.0",
    "s3fs>=0.6.0",
    "polars>=1.8.2",
    "duckdb>=1.1.3",
    "pyarrow>=17.0.0",
    "marimo>=0.8.22",
    "matplotlib>=3.7.5",
    "seaborn>=0.13.2",
    "scipy>=1.10.1",
    "requests>=2.32.4",
    "adlfs>=2024.7.0",
    "sqlalchemy",
    "alembic",
    "pydantic",
    "python-dotenv",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "PyJWT>=2.8.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "psycopg[binary]",
    "chromadb>=0.5.0",
    "sentence-transformers>=2.2.0",
    "onnxruntime==1.16.3; python_version < '3.10'",
    "onnxruntime==1.20.1; python_version >= '3.10'",
    "mlflow>=2.0.0",
    "cornac>=2.3.0",
    "implicit>=0.7.2",
    "python-dotenv>=1.0.0",
    "boto3>=1.37.3",
    "sentence-transformers>=3.2.1",
    "torch>=2.5.1",
    "datasets>=2.14.4",
    "accelerate>=1.0.1",
    "transformers[torch]>=4.46.3",
    "duckdb>=1.3.2",
    "sqlglot>=26.33.0",
    "unsloth>=2026.1.2",
]

[project.optional-dependencies]
unsloth = [
    "unsloth>=2026.1.1; platform_system == 'Linux' and python_version >= '3.9' and python_version < '3.14'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "bookdb/__init__.py"
[tool.hatch.build.targets.sdist]
include = [
    "/bookdb",
]

[tool.hatch.build.targets.wheel]
packages = ["bookdb"]

[dependency-groups]
dev = [
    "black>=22.0.0",
    "pytest>=7.0.0",
]
